find_topic and rocon_master_info finalised.
