+++ rocon_python_comms/src/rocon_python_comms/__init__.py
@@ -11,4 +11,5 @@ from .service_pair_client import ServicePairClient
 from .service_pair_server import ServicePairServer
 from .services import find_service
 from .subscriber_proxy import SubscriberProxy
+from .topics import find_topic
 from .wall_rate import WallRate
