+++ rocon_python_comms/src/rocon_python_comms/services.py
@@ -8,8 +8,9 @@
 
 import rospy
 import time
-import rostopic
+from rosservice import rosservice_find, ROSServiceIOException
 
+# Local imports
 from .exceptions import NotFoundException
 
 ##############################################################################
@@ -17,14 +18,14 @@ from .exceptions import NotFoundException
 ##############################################################################
 
 
-def find_topic(topic_type, timeout=rospy.rostime.Duration(5.0), unique=False):
+def find_service(service_type, timeout=rospy.rostime.Duration(5.0), unique=False):
     '''
-      Do a lookup to find topics of the type
+      Do a lookup to find services of the type
       specified. This will raise exceptions if it times out or returns
       multiple values. It can apply the additional logic of whether this should
       return a single unique result, or a list.
 
-      @param topic_type : topic type specification, e.g. rocon_std_msgs/MasterInfo
+      @param service_type : service type specification, e.g. concert_msgs/GetInteractions
       @type str
 
       @param timeout : raise an exception if nothing is found before this timeout occurs.
@@ -33,26 +34,26 @@ def find_topic(topic_type, timeout=rospy.rostime.Duration(5.0), unique=False):
       @param unique : flag to select the lookup behaviour (single/multiple results)
       @type bool
 
-      @return the fully resolved name of the topic (unique) or list of names (non-unique)
+      @return the fully resolved name of the service (unique) or list of names (non-unique)
       @type str
 
       @raise rocon_python_comms.NotFoundException
     '''
-    topic_name = None
-    topic_names = []
+    service_name = None
+    service_names = []
     timeout_time = time.time() + timeout.to_sec()
-    while not rospy.is_shutdown() and time.time() < timeout_time and not topic_names:
+    while not rospy.is_shutdown() and time.time() < timeout_time and not service_names:
         try:
-            topic_names = rostopic.find_by_type(topic_type)
+            service_names = rosservice_find(service_type)
         except ROSServiceIOException:
             raise NotFoundException("ros shutdown")
         if unique:
-            if len(topic_names) > 1:
-                raise NotFoundException("multiple services found %s." % topic_names)
-            elif len(topic_names) == 1:
-                topic_name = topic_names[0]
-        if not topic_names:
+            if len(service_names) > 1:
+                raise NotFoundException("multiple services found %s." % service_names)
+            elif len(service_names) == 1:
+                service_name = service_names[0]
+        if not service_names:
             rospy.rostime.wallsleep(0.1)
-    if topic_name is None:
+    if service_name is None:
         raise NotFoundException("timed out")
-    return topic_name
+    return service_name
