+++ rocon_python_comms/src/rocon_python_comms/topics.py
@@ -8,9 +8,8 @@
 
 import rospy
 import time
-from rosservice import rosservice_find, ROSServiceIOException
+import rostopic
 
-# Local imports
 from .exceptions import NotFoundException
 
 ##############################################################################
@@ -18,14 +17,14 @@ from .exceptions import NotFoundException
 ##############################################################################
 
 
-def find_service(service_type, timeout=rospy.rostime.Duration(5.0), unique=False):
+def find_topic(topic_type, timeout=rospy.rostime.Duration(5.0), unique=False):
     '''
-      Do a lookup to find services of the type
+      Do a lookup to find topics of the type
       specified. This will raise exceptions if it times out or returns
       multiple values. It can apply the additional logic of whether this should
       return a single unique result, or a list.
 
-      @param service_type : service type specification, e.g. concert_msgs/GetInteractions
+      @param topic_type : topic type specification, e.g. rocon_std_msgs/MasterInfo
       @type str
 
       @param timeout : raise an exception if nothing is found before this timeout occurs.
@@ -34,26 +33,26 @@ def find_service(service_type, timeout=rospy.rostime.Duration(5.0), unique=False
       @param unique : flag to select the lookup behaviour (single/multiple results)
       @type bool
 
-      @return the fully resolved name of the service (unique) or list of names (non-unique)
+      @return the fully resolved name of the topic (unique) or list of names (non-unique)
       @type str
 
       @raise rocon_python_comms.NotFoundException
     '''
-    service_name = None
-    service_names = []
+    topic_name = None
+    topic_names = []
     timeout_time = time.time() + timeout.to_sec()
-    while not rospy.is_shutdown() and time.time() < timeout_time and not service_names:
+    while not rospy.is_shutdown() and time.time() < timeout_time and not topic_names:
         try:
-            service_names = rosservice_find(service_type)
-        except ROSServiceIOException:
+            topic_names = rostopic.find_by_type(topic_type)
+        except rostopic.ROSTopicException:
             raise NotFoundException("ros shutdown")
         if unique:
-            if len(service_names) > 1:
-                raise NotFoundException("multiple services found %s." % service_names)
-            elif len(service_names) == 1:
-                service_name = service_names[0]
-        if not service_names:
+            if len(topic_names) > 1:
+                raise NotFoundException("multiple topics found %s." % topic_names)
+            elif len(topic_names) == 1:
+                topic_name = topic_names[0]
+        if not topic_names:
             rospy.rostime.wallsleep(0.1)
-    if service_name is None:
+    if topic_name is None:
         raise NotFoundException("timed out")
-    return service_name
+    return topic_name
